requirements.txt
pynacl
yt-dlp
dotenv
pytube
discord.py
spotipy
ffmpeg